writing non-ASCII data
I found that subprocess.jsm does correctly write 8-bit data (binary,
UTF-8, etc.). The main problem is that ctypes.char does a conversion of
the input data to UTF-8, whether or not this is desired.

The problems are:
- true UTF-8 is truncated because the length of the input string
(Unicode) is shorter than the length of the string passed to write/WriteFile
- binary data and data in other charsets get garbled.

I have attached a solution for this. The key point is to not use
ctypes.char.ptr but ctypes.uint_8.array.

The disadvantage is that the user needs to convert the string before
calling stdin.write() to the correct encoding. However, I tried XPCOM
objects like Ci.nsIFileOutputStream and they behave the same way.

This means that the "charset" option is only relevant to data *read*
from the subprocess, but not to the data written to the subprocess.